Change selector to a CSS-style selector
With `querySelectorAll`, CSS-style selectors are possible, which are more flexible than a class selector. At least in my use-case, this means I can use it in a Jekyll generated site where I cannot easily change the internal structure of the code samples and add the `tei-doc-link` class. I can then set `TEI_DOC_LINK` to `pre.highlight code`. 

Afaik there are no downsides, the only caveat: the list returned by `querySelectorAll()` is not live, as opposed to `getElementsByClassName`. (https://stackoverflow.com/a/26757629) You might if that poses a problem (it doesn’t where I want to use it).